+++ plugins/org.preesm.ui.pisdf/src/org/preesm/ui/pisdf/features/SetActorRefinementFeature.java
@@ -241,13 +241,15 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
 
         // We get it
         file = ResourcesPlugin.getWorkspace().getRoot().getFile(newFilePath);
+        if (file == null) {
+          throw new PreesmRuntimeException("Unable to open file " + newFilePath.toString());
+        }
 
         // Get all prototypes first (no filter)
-        final List<FunctionPrototype> allPrototypes = getPrototypes(file, actor, PrototypeFilter.NONE);
+        final List<FunctionPrototype> allPrototypes = HeaderParser.parseCXXHeader(file);
         allProtoArray = allPrototypes.toArray(new FunctionPrototype[allPrototypes.size()]);
 
-        loopPrototypes = getPrototypes(file, actor, PrototypeFilter.LOOP_ACTOR);
-        validRefinement = (!loopPrototypes.isEmpty()) || (!allPrototypes.isEmpty());
+        validRefinement = !allPrototypes.isEmpty();
         if (!validRefinement) {
           final String message = "The .h file you selected does not contain any prototype."
               + ".\nPlease select another valid file.";
@@ -268,6 +270,8 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
             title = "Loop Function Selection";
             message = "Select a loop function for actor " + ((AbstractActor) actor).getName()
                 + "\n(* = any string, ? = any char):";
+
+            loopPrototypes = getPrototypes(actor, PrototypeFilter.LOOP_ACTOR, allPrototypes);
             final FunctionPrototype[] loopProtoArray = loopPrototypes
                 .toArray(new FunctionPrototype[loopPrototypes.size()]);
             loopProto = PiMMUtil.selectFunction(loopProtoArray, allProtoArray, title, message,
@@ -278,10 +282,10 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
           List<FunctionPrototype> allInitPrototypes = null;
           boolean showOnlyCorresponding = false;
           if (actor instanceof Actor) {
-            initPrototypes = getPrototypes(file, actor, PrototypeFilter.INIT_ACTOR);
-            allInitPrototypes = getPrototypes(file, actor, PrototypeFilter.INIT);
+            initPrototypes = getPrototypes(actor, PrototypeFilter.INIT_ACTOR, allPrototypes);
+            allInitPrototypes = getPrototypes(actor, PrototypeFilter.INIT, allPrototypes);
           } else {
-            initPrototypes = getPrototypes(file, actor, PrototypeFilter.INIT_DELAY_ACTOR);
+            initPrototypes = getPrototypes(actor, PrototypeFilter.INIT_DELAY_ACTOR, allPrototypes);
             allInitPrototypes = new ArrayList<>();
             showOnlyCorresponding = true;
           }
@@ -329,40 +333,37 @@ public class SetActorRefinementFeature extends AbstractCustomFeature {
   /**
    * Gets the prototypes.
    *
-   * @param file
-   *          the file
    * @param actor
    *          the actor
    * @param prototypeFilter
    *          the prototype filter
+   * @param allPrototypes
+   *          All the prototypes found in the file
    * @return the prototypes
    */
-  private List<FunctionPrototype> getPrototypes(final IFile file, final RefinementContainer actor,
-      final PrototypeFilter prototypeFilter) {
+  private List<FunctionPrototype> getPrototypes(final RefinementContainer actor, final PrototypeFilter prototypeFilter,
+      final List<FunctionPrototype> allPrototypes) {
 
     List<FunctionPrototype> result = null;
 
-    if (file != null) {
-      result = HeaderParser.parseHeader(file);
-
-      switch (prototypeFilter) {
-        case INIT_ACTOR:
-          result = HeaderParser.filterInitPrototypesFor((AbstractActor) actor, result);
-          break;
-        case LOOP_ACTOR:
-          result = HeaderParser.filterLoopPrototypesFor((AbstractActor) actor, result);
-          break;
-        case INIT:
-          result = HeaderParser.filterInitPrototypes(result);
-          break;
-        case INIT_DELAY_ACTOR:
-          result = HeaderParser.filterInitBufferPrototypes(result);
-          break;
-        case NONE:
-          break;
-        default:
-      }
+    switch (prototypeFilter) {
+      case INIT_ACTOR:
+        result = HeaderParser.filterInitPrototypesFor((AbstractActor) actor, allPrototypes);
+        break;
+      case LOOP_ACTOR:
+        result = HeaderParser.filterLoopPrototypesFor((AbstractActor) actor, allPrototypes);
+        break;
+      case INIT:
+        result = HeaderParser.filterInitPrototypes(allPrototypes);
+        break;
+      case INIT_DELAY_ACTOR:
+        result = HeaderParser.filterInitBufferPrototypes(allPrototypes);
+        break;
+      case NONE:
+        break;
+      default:
     }
+
     if (result == null) {
       throw new PreesmRuntimeException();
     }